+++ Trappist/src/Promact.Trappist.DomainModel/Models/Enums/DifficultyLevel.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace Promact.Trappist.DomainModel.Models.Enums
+{
+    public enum DifficultyLevel
+    {
+        Easy, Medium, Hard
+    }
+}
