--- Trappist/src/Promact.Trappist.DomainModel/Models/Enums/EnumList.cs
@@ -1,23 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace Promact.Trappist.DomainModel.Models.Enums
-{
-    public class EnumList
-    {
-        public enum QuestionType
-        {
-            Single, Multiple, Programming
-        }
-        public enum DifficultyLevel
-        {
-            Easy, Medium, Hard
-        }
-        
-    }
-   
-
-
-}
