+++ Trappist/src/Promact.Trappist.DomainModel/Models/Question/SingleMultipleAnswerQuestion.cs
@@ -6,7 +6,7 @@ namespace Promact.Trappist.DomainModel.Models.Question
     public class SingleMultipleAnswerQuestion : QuestionBase
     {
         [Required]
-        public ICollection<SingleMultipleAnswerQuestionOption> SingleMutipleAnswerQuestionOption { get; set; }
+        public virtual ICollection<SingleMultipleAnswerQuestionOption> SingleMutipleAnswerQuestionOption { get; set; }
 
     }
 }
