+++ Trappist/src/Promact.Trappist.DomainModel/Models/Question/SingleMultipleAnswerQuestionOption.cs
@@ -1,14 +1,14 @@
 ﻿using System.ComponentModel.DataAnnotations;
-using System.ComponentModel.DataAnnotations.Schema;
 namespace Promact.Trappist.DomainModel.Models.Question
 {
     public class SingleMultipleAnswerQuestionOption:BaseModel
     {
         [Required]
         public string Option { get; set; }
+        [Required]
         public bool IsAnswer { get; set; }
-        [ForeignKey("Question_Id")]
-        public SingleMultipleAnswerQuestion SingleMultipleAnswerQuestion { get; set; }
+        public int SingleMultipleAnswerQuestionID { get; set; }
+        public virtual SingleMultipleAnswerQuestion SingleMultipleAnswerQuestion { get; set; }
 
     }
 }
